Fix to not mangle existing properties on files (#20)
* If file has properties the "top of the note" will now be the first line after the properties.
This prevents properties from being mangled when indexing a note. Espeically important with the advent of the new Bases core plugin.
Added in a test case as well.

* Had to account for horizontal rules further down a file so that they didn't get picked up as "properties"

Properties will not be recognized if the triple dash is not at the very first line which matches how obsidian handles properties.

* Change +=1 to ++ for countTripDash
Run Prettier on changed files index.svc and index.spec.ts.

---------

Co-authored-by: Bryce Leo <bryce.leo@trainingvault.us>